refactor(#152) :: admin/feedid에 pageable추가

+++ src/main/java/com/command/toyvillage_server/domain/app/feed_log/presentation/dto/response/FeedLogAdminQueryListResponse.java
@@ -23,7 +23,7 @@ public record FeedLogAdminQueryListResponse(
             String animalKind,
             String animalName,
             String feedType,
-            Integer feedAmount,
+            Float feedAmount,
             LocalDateTime feedDateTime
     ) {
         public static FeedLogResponse from(FeedLog feedLog) {
